requirements.txt
APScheduler==3.6.0
astral==1.10.1
btlewrap==0.0.4
certifi==2019.3.9
chardet==3.0.4
colorize==1.0.4
cycler==0.10.0
giphypop==0.3
idna==2.8
kiwisolver==1.0.1
matplotlib==3.0.3
miflora==0.5
numpy==1.16.2
pandas==0.24.2
Pillow==6.0.0
PTable==0.9.2
pyparsing==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
slackclient==1.3.1
tzlocal==1.5.1
urllib3==1.24.1
weather-api==1.0.7
websocket-client==0.54.0
